Move dbs in submodules
